fides/requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-migrate>=4.0
werkzeug>=3.0
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.0
resend>=0.7
anthropic>=0.20
requests>=2.31
beautifulsoup4>=4.12
feedparser>=6.0
